Add method for updating bots
This commit adds bot an update page and get/post methods to update bots
on the server. It allows for editing the Dockerfile using a <textarea>
filed and also uploading(replacing) the zip file.

This change concludes all the core features needed for the service to
work. After removing the admin password and the auto-fill added for
development, this service can be deployed